# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AT Réservations API
  description: |
    API REST pour la gestion des missions et réservations — Algérie Télécom.

    ## Authentification
    L'API utilise **Laravel Sanctum** avec tokens httpOnly cookie.
    Appelez `POST /auth/login` pour obtenir un cookie de session,
    puis incluez-le automatiquement dans les requêtes suivantes.

    ## Rôles
    | Rôle | Description |
    |------|-------------|
    | `admin` | Administrateur système — accès total |
    | `validateur` | Valide/rejette les missions soumises |
    | `utilisateur` | Crée et gère ses propres missions |
    | `demandeur` | Soumet des demandes de mission |

    ## Cycle de vie d'une mission
    `brouillon` → `soumis` → `en_validation` → `approuve` / `rejete` / `annule`
  version: 1.0.0
  contact:
    name: Support IT Algérie Télécom
    email: it-support@at.dz
servers:
  - url: https://backend-production-170c.up.railway.app/api
    description: Production (Railway)
  - url: http://localhost:8000/api
    description: Développement local
...
components:
  securitySchemes:
    sanctum:
      type: apiKey
      in: cookie
      name: at_reservations_token
      description: Cookie httpOnly Sanctum
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
...
    Mission:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        numero_unique:
          type: string
          example: "OM-2026-0042"
        titre:
          type: string
        objet_mission:
          type: string
        destination:
          type: string
          nullable: true
        destination_ville:
          type: string
        date_depart:
          type: string
          format: date
        date_retour:
          type: string
          format: date
        type_mission:
          type: string
          enum: [nationale, internationale, formation, conference, reunion, inspection, audit, autre]
        priorite:
          type: string
          enum: [normale, urgente, tres_urgente]
        statut:
          type: string
          enum: [brouillon, soumis, en_validation, approuve, rejete, annule]
        budget_previsionnel:
          type: number
          format: float
          nullable: true
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
...
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
...
security:
  - sanctum: []
paths:
...
  /missions/{id}:
    get:
      tags: [Missions]
      summary: Détail d'une mission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détail mission
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Mission'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Missions]
      summary: Modifier une mission (brouillon uniquement)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                titre:
                  type: string
                objet_mission:
                  type: string
                destination_ville:
                  type: string
                date_depart:
                  type: string
                  format: date
                date_retour:
                  type: string
                  format: date
      responses:
        '200':
          description: Mission mise à jour
        '403':
          description: Modification interdite (statut non brouillon)

    delete:
      tags: [Missions]
      summary: Supprimer une mission (brouillon uniquement)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Mission supprimée
        '403':
          description: Suppression interdite
...